Integration of garmin sync admin panel Integration of autoposting admin panel (mid way)

+++ class/ActivityFile.php
@@ -13,6 +13,7 @@ class ActivityFile extends Model {
     public $filename;
     public $ext;
     public $data;
+    public $latest_error = null;
     public $allowed_extensions = ['gpx', 'fit'];
 
     function __construct ($id = NULL) {
@@ -21,9 +22,11 @@ class ActivityFile extends Model {
         $data = $this->getData($this->table);
         if ($id != NULL) {
             $this->user_id = $data['user_id'];
+            $this->activity_id = $data['activity_id'];
             $this->filename = $data['filename'];
             $this->ext = $data['ext'];
-            $this->posting_date = (new DateTime($data['posting_date'], new DateTimezone('Asia/Tokyo')));
+            $this->posting_date = (new DateTime($data['posting_date']))->setTimezone(new DateTimeZone('Asia/Tokyo'))->format('Y-m-d H:i:s');
+            $this->latest_error = $data['latest_error'];
         }
     }
 
@@ -47,7 +50,8 @@ class ActivityFile extends Model {
         $this->user_id = $metadata['user_id'];
         $this->ext = $metadata['ext'];
         $this->filename = setFilename('activity', $this->ext);
-        $this->posting_date = (new DateTime(date('Y-m-d H:i:s'), new DateTimezone('Asia/Tokyo')));
+        $this->posting_date = (new DateTime(date('Y-m-d H:i:s')))->setTimeZone('Asia/Tokyo');
+        $this->id = getNextAutoIncrement($this->table);
 
         // Insert file data in table
         $insertFileData = $this->getPdo()->prepare('INSERT INTO activity_files(user_id, activity_id, filename, ext, posting_date) VALUES (?, ?, ?, ?, ?)');
@@ -92,9 +96,12 @@ class ActivityFile extends Model {
 
             // On success, send analyzed data back to client
             $phpGpx = new phpGPX();
-            $gpx = $phpGpx->load($this->getUrl());
+            if (file_exists($this->getUrl())) $gpx = $phpGpx->load($this->getUrl());
+            else throw new Exception('file_not_found');
             $activity_data = new ActivityData();
             $activity_data->buildFromGpx($gpx);
+            $activity_data->file_id = $this->id;
+
             $this->data = $activity_data;
             return $activity_data;
         
@@ -107,7 +114,8 @@ class ActivityFile extends Model {
             if (!is_dir($_SERVER["DOCUMENT_ROOT"] . '/media/activities/data')) mkdir($_SERVER["DOCUMENT_ROOT"] . '/media/activities/data');
             if (!is_dir($_SERVER["DOCUMENT_ROOT"] . '/media/activities/data/temp')) mkdir($_SERVER["DOCUMENT_ROOT"] . '/media/activities/data/temp');
             $url = $_SERVER["DOCUMENT_ROOT"] . '/media/activities/data/temp/temp.' . $this->ext;
-            file_put_contents($url, file_get_contents($this->getUrl()));
+            if (file_exists($this->getUrl())) file_put_contents($url, file_get_contents($this->getUrl()));
+            else throw new Exception('file_not_found');
 
             // Then parse
 
@@ -116,6 +124,7 @@ class ActivityFile extends Model {
             
             $activity_data = new ActivityData();
             $activity_data->buildFromFit($fit);
+            $activity_data->file_id = $this->id;
 
             $this->data = $activity_data;
             return $activity_data;
